move operation codes enum out of phase 2 folder
