+++ packages/core/lib/triggers/TRPS0011.test.ts
@@ -2,7 +2,7 @@ import disarrCompatibleResultClass from "../../phase2/lib/getOperationSequence/g
 import isRecordableOffence from "../../phase2/lib/isRecordableOffence"
 import generateAhoFromOffenceList from "../../phase2/tests/fixtures/helpers/generateAhoFromOffenceList"
 import generatePncUpdateDatasetFromOffenceList from "../../phase2/tests/fixtures/helpers/generatePncUpdateDatasetFromOffenceList"
-import { PNCMessageType } from "../../phase2/types/operationCodes"
+import { PNCMessageType } from "../../types/operationCodes"
 import type { Offence } from "../../types/AnnotatedHearingOutcome"
 import Phase from "../../types/Phase"
 import hasCompletedDisarr from "./hasCompletedDisarr"
