+++ packages/core/phase2/lib/getOperationSequence/generateOperations/createOperation.test.ts
@@ -1,4 +1,4 @@
-import { PNCMessageType } from "../../../types/operationCodes"
+import { PNCMessageType } from "../../../../types/operationCodes"
 import createOperation from "./createOperation"
 
 describe("createOperation", () => {
