+++ packages/core/phase2/lib/getOperationSequence/generateOperations/filterDisposalsAddedInCourt.ts
@@ -1,5 +1,5 @@
 import type { Operation } from "../../../../types/PncUpdateDataset"
-import { PNCMessageType } from "../../../types/operationCodes"
+import { PNCMessageType } from "../../../../types/operationCodes"
 import extractRemandCcrs from "./extractRemandCcrs"
 
 const filterDisposalsAddedInCourt = (operations: Operation[]): Operation[] => {
