+++ packages/core/phase2/lib/getOperationSequence/generateOperations/resultClassHandlers/handleAdjournmentPostJudgement.test.ts
@@ -1,6 +1,6 @@
 import type { Offence, Result } from "../../../../../types/AnnotatedHearingOutcome"
 import generateResultClassHandlerParams from "../../../../tests/helpers/generateResultClassHandlerParams"
-import { PNCMessageType } from "../../../../types/operationCodes"
+import { PNCMessageType } from "../../../../../types/operationCodes"
 import { handleAdjournmentPostJudgement } from "./handleAdjournmentPostJudgement"
 
 const organisationUnit = {
