+++ packages/core/phase2/lib/getOperationSequence/generateOperations/resultClassHandlers/handleAdjournmentPreJudgement.test.ts
@@ -1,7 +1,7 @@
 import type { Result } from "../../../../../types/AnnotatedHearingOutcome"
 import ResultClass from "../../../../../types/ResultClass"
 import generateResultClassHandlerParams from "../../../../tests/helpers/generateResultClassHandlerParams"
-import { PNCMessageType } from "../../../../types/operationCodes"
+import { PNCMessageType } from "../../../../../types/operationCodes"
 import { handleAdjournmentPreJudgement } from "./handleAdjournmentPreJudgement"
 
 const organisationUnit = {
