+++ packages/core/phase2/lib/getOperationSequence/generateOperations/resultClassHandlers/handleAppealOutcome.ts
@@ -2,7 +2,7 @@ import ExceptionCode from "bichard7-next-data-latest/dist/types/ExceptionCode"
 import errorPaths from "../../../../../lib/exceptions/errorPaths"
 import createOperation from "../createOperation"
 import type { ResultClassHandler } from "./ResultClassHandler"
-import { PNCMessageType } from "../../../../types/operationCodes"
+import { PNCMessageType } from "../../../../../types/operationCodes"
 
 export const handleAppealOutcome: ResultClassHandler = ({ result, offence, offenceIndex, resultIndex }) => {
   if (result.PNCAdjudicationExists) {
