+++ packages/core/phase2/lib/getOperationSequence/operationCourtCaseReference.ts
@@ -1,5 +1,5 @@
 import type { Operation } from "../../../types/PncUpdateDataset"
-import { PNCMessageType } from "../../types/operationCodes"
+import { PNCMessageType } from "../../../types/operationCodes"
 
 const operationCodes = [
   PNCMessageType.SENTENCE_DEFERRED,
