+++ packages/core/phase2/lib/getOperationSequence/sortOperations.test.ts
@@ -1,5 +1,5 @@
 import type { Operation } from "../../../types/PncUpdateDataset"
-import { PNCMessageType } from "../../types/operationCodes"
+import { PNCMessageType } from "../../../types/operationCodes"
 import sortOperations from "./sortOperations"
 
 describe("check sortOperations", () => {
