+++ packages/core/phase2/lib/getOperationSequence/validateOperations.test.ts
@@ -1,5 +1,5 @@
 import type { Operation } from "../../../types/PncUpdateDataset"
-import { PNCMessageType } from "../../types/operationCodes"
+import { PNCMessageType } from "../../../types/operationCodes"
 import validateOperations from "./validateOperations"
 
 const allOperations = [
